feat: add index file that imports bits from all downstream files (#15)
This means that the entrypoint that downstream code needs to import
from can be a fixed / predictable file and not a file that changes name.

I considered just renaming the `-node.ts` file but the reason I made
this name variable was so that I could test multiple different packages
dumping in the same output directory and that is kinda a nice property
that I'd like to keep working at least for the time being.